requirements.txt
asgiref==3.4.1
autopep8==1.6.0
dj-database-url==0.5.0
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.4.0
psycopg2==2.9.2
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
whitenoise==5.3.0
